+++ src/main/java/com/github/zeykrus/bankpet/exception/InsufficientFundsException.java
@@ -1,14 +1,30 @@
 package com.github.zeykrus.bankpet.exception;
 
+/**
+ * Исключение, выбрасываемое при попытке снять сумму,
+ * превышающую доступный баланс (с учётом кредитного лимита).
+ *
+ */
 public class InsufficientFundsException extends Exception {
-    private final double deficit;
+    private final long deficit;
 
-    public InsufficientFundsException(String message, double deficit) {
+    /**
+     * Создаёт исключение с сообщением и суммой недостающих средств.
+     *
+     * @param message сообщение об ошибке
+     * @param deficit сумма, которой не хватает
+     */
+    public InsufficientFundsException(String message, long deficit) {
         super(message);
         this.deficit = deficit;
     }
 
-    public double getDeficit() {
+    /**
+     * Возвращает сумму недостающих средств.
+     *
+     * @return дефицит средств
+     */
+    public long getDeficit() {
         return deficit;
     }
 }
